requirements.txt
pandas==2.1.4
requests==2.31.0
google-cloud-bigquery==3.14.1
flask==3.0.0
gunicorn==21.2.0